refactor: Refactored organization invite template from invitation-email template
